cast is_eop

+++ function_builder.py
@@ -157,7 +157,7 @@ def get_classification_loss(
   seg_id = tf.transpose(features["segment_ids"], [1, 0])
   inp_mask = tf.transpose(features["input_mask"], [1, 0])
   label = tf.reshape(features["label_ids"], [bsz_per_core])
-  is_eop = features["is_eop"]
+  is_eop = tf.cast(features["is_eop"], dtype=tf.float32)
 
   xlnet_config = xlnet.XLNetConfig(json_path=FLAGS.model_config_path)
   run_config = xlnet.create_run_config(is_training, True, FLAGS)
